Fix building on MinGW again (hope I didn't break MSVC build)

+++ src/r_things.cc
@@ -99,7 +99,7 @@ static const image_c *shadow_image = NULL;
 static float GetHoverDZ(mobj_t *mo)
 {
 	// compute a different phase for different objects
-	angle_t phase = (angle_t)(long)mo; //x64: Removed C4311 warning (right part of expression is now 'long long' instead of just 'long'.
+	angle_t phase = (angle_t)(long)(intptr_t)mo; //x64: Removed C4311 warning (right part of expression is now 'long long' instead of just 'long'.
 	phase ^= (angle_t)(phase << 19);
 	phase += (angle_t)(leveltime << (ANGLEBITS-6));
